+++ app/api/utils/contract_actions/rumble_kong_league.py
@@ -1,5 +1,6 @@
 
 from .contract_instance import get_contract_instance
+from .contract_actions import (get_token_ids, get_token_uri)
 
 
 def get_rkl_address_token_count(address):
@@ -8,3 +9,15 @@ def get_rkl_address_token_count(address):
     """
     contract = get_contract_instance('rkl')
     return contract.functions.balanceOf(address).call()
+
+
+def get_rkl_tokens(address) -> list:
+    contract = get_contract_instance('rkl')
+    token_ids = get_token_ids(address, contract)
+    tokens = []
+    for token_id in token_ids:
+        token = get_token_uri(token_id, contract)
+        if token:
+            token['token_id'] = token_id
+            tokens.append(token)
+    return tokens
