+++ nginx/nginx.Dockerfile
@@ -1,0 +1,3 @@
+FROM nginx:latest
+
+COPY ./nginx/nginx.conf /etc/nginx/conf.d/default.conf
