+++ nginx/nginx.conf
@@ -1,0 +1,38 @@
+upstream app {
+	server app:8000;
+}
+
+server {
+
+	client_max_body_size 4G;
+	keepalive_timeout 5;
+
+	location / {
+		try_files $uri @proxy_to_app;
+	}
+
+	location @proxy_to_app {
+
+		proxy_pass http://app;
+
+		# proxy_http_version 1.1;
+		# proxy_set_header Upgrade $http_upgrade;
+		# proxy_set_header Connection "upgrade";
+
+		proxy_redirect off;
+		proxy_set_header Host $host;
+		proxy_set_header X-Real-IP $remote_addr;
+		proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+		proxy_set_header X-Forwarded-Host $server_name;
+	}
+	location /backendstatic/ {
+		root /synthesis/;
+	}
+
+	location /favicon.ico {
+		access_log off; log_not_found off;
+	}
+	location /static/ {
+	}
+
+}
